praas-app: rename pages/conduit to pages/conduits
update conduit route as conduits
